org groups table role cell, set or update role
